move to lib and integrate into generate.js
